+++ activerecord/lib/active_record/connection_adapters/postgresql/cast.rb
@@ -6,8 +6,8 @@ module ActiveRecord
           return string unless String === string
 
           case string
-          when 'infinity'; 1.0 / 0.0
-          when '-infinity'; -1.0 / 0.0
+          when 'infinity'; ::Float::INFINITY
+          when '-infinity'; -::Float::INFINITY
           else
             super
           end
@@ -58,6 +58,12 @@ module ActiveRecord
             end
           end
           "{#{casted_values.join(',')}}"
+        end
+
+        def range_to_string(object)
+          from = object.begin.respond_to?(:infinite?) && object.begin.infinite? ? '' : object.begin
+          to   = object.end.respond_to?(:infinite?) && object.end.infinite? ? '' : object.end
+          "[#{from},#{to}#{object.exclude_end? ? ')' : ']'}"
         end
 
         def string_to_json(string)